+++ .github/workflows/ci.yml
@@ -16,7 +16,7 @@ jobs:
     - name: Set up Go
       uses: actions/setup-go@v5.3.0
       with:
-        go-version: '1.24.2'
+        go-version: 'stable'
 
     - name: Build
       run: ./tools/build.sh
@@ -29,7 +29,7 @@ jobs:
     - name: Set up Go
       uses: actions/setup-go@v5.3.0
       with:
-        go-version: '1.24.2'
+        go-version: 'stable'
 
     - name: Test
       run: go test -race ./...
@@ -41,7 +41,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions/setup-go@v5.3.0
         with:
-          go-version: '1.24.2'
+          go-version: 'stable'
       - name: golangci-lint
         uses: golangci/golangci-lint-action@v7.0.0
         with:
