+++ go.mod
@@ -3,15 +3,16 @@ module github.com/Lexer747/acci-ping
 go 1.24
 
 require (
-	golang.org/x/exp v0.0.0-20250305212735-054e65f0b394
-	golang.org/x/net v0.39.0
-	golang.org/x/term v0.31.0
+	golang.org/x/exp v0.0.0-20250506013437-ce4c2cf36ca6
+	golang.org/x/net v0.40.0
+	golang.org/x/term v0.32.0
 )
 
 // Test dependencies
 require (
 	github.com/google/go-cmp v0.7.0
 	gotest.tools/v3 v3.5.2
+	pgregory.net/rapid v1.2.0
 )
 
-require golang.org/x/sys v0.32.0 // indirect
+require golang.org/x/sys v0.33.0 // indirect
