+++ graph/data/longer_test.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+go test -rapid.checks=10_000
