--- graph/data/testdata/.gitignore
@@ -1,2 +1,0 @@
-# Test outputs which are not GOLDEN's
-*.actual
