+++ gui/box_test/hashes
@@ -1,0 +1,2 @@
+Hash Count: 4860
+Hash XOR: gzvkMNkVVIRh7SlpRVWKYK02YmBHA4N2/iWj3XFbp/w=
