+++ tools/linecount.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+# shellcheck disable=SC2046
+# shellcheck disable=SC2006
+wc -l `find ./ -type f -name '*.go'` | sort
